frontend/requirements.txt
requests==2.32.3
streamlit==1.36.0
python-dotenv==1.0.1
pysqlite3-binary
langchain-groq
duckduckgo-search
langchain
langchain-community 
langchainhub 
tiktoken 
langgraph
langchain-chroma
bs4
huggingface_hub 
unstructured 
sentence_transformers
chromadb
beautifulsoup4==4.12.3
